test(tokenless): raise test coverage from 59% to 75%
Add 100+ unit tests and 18 CLI integration tests across all four
tokenless crates. Test code extracted into separate files under
src/tests/ via include!() for cleaner source separation.

Remaining uncovered code is in I/O-bound CLI dispatch functions
(run(), auto_fix(), serve()) that cannot be instrumented by
tarpaulin when invoked via subprocess.
