+++ src/tokenless/crates/tokenless-ccr/src/tests/store_tests.rs
@@ -1,0 +1,15 @@
+use crate::InMemoryStore;
+
+#[test]
+fn is_empty_default_trait_method() {
+    let store = InMemoryStore::new();
+    assert!(store.is_empty());
+    store.stash("payload").unwrap();
+    assert!(!store.is_empty());
+}
+
+#[test]
+fn stash_error_display() {
+    let e = StashError::Backend("test error".to_string());
+    assert!(format!("{}", e).contains("test error"));
+}
