+++ src/tokenless/crates/tokenless-cli/src/env_check.rs
@@ -1321,453 +1321,5 @@ pub fn run(
 #[cfg(test)]
 mod tests {
     use super::*;
-    use serde_json::json;
-
-    #[test]
-    fn normalize_dep_simple_string() {
-        let dep = normalize_dep(&json!("jq"));
-        assert_eq!(dep.binary, "jq");
-        assert_eq!(dep.package, "jq");
-        assert_eq!(dep.manager, "rpm");
-        assert!(dep.version.is_none());
-        assert!(dep.fallback.is_empty());
-    }
-
-    #[test]
-    fn normalize_dep_version_string() {
-        let dep = normalize_dep(&json!("rtk>=0.35"));
-        assert_eq!(dep.binary, "rtk");
-        assert_eq!(dep.version.as_deref(), Some(">=0.35"));
-        assert_eq!(dep.package, "rtk");
-        assert_eq!(dep.manager, "rpm");
-    }
-
-    #[test]
-    fn normalize_dep_object() {
-        let dep = normalize_dep(&json!({
-            "binary": "curl",
-            "package": "curl",
-            "manager": "rpm"
-        }));
-        assert_eq!(dep.binary, "curl");
-        assert_eq!(dep.package, "curl");
-        assert_eq!(dep.manager, "rpm");
-        assert!(dep.version.is_none());
-    }
-
-    #[test]
-    fn normalize_dep_object_with_all_fields() {
-        let dep = normalize_dep(&json!({
-            "binary": "rtk",
-            "version": ">=0.35",
-            "package": "rtk",
-            "manager": "cargo",
-            "pip_name": "rtk-pip",
-            "uv_name": "rtk-uv",
-            "npm_name": "rtk-npm",
-            "use_npx": true,
-            "fallback": [
-                {"method": "symlink", "binary": "rtk", "source": "/usr/libexec/anolisa/tokenless/rtk"}
-            ]
-        }));
-        assert_eq!(dep.binary, "rtk");
-        assert_eq!(dep.version.as_deref(), Some(">=0.35"));
-        assert_eq!(dep.manager, "cargo");
-        assert_eq!(dep.pip_name.as_deref(), Some("rtk-pip"));
-        assert_eq!(dep.uv_name.as_deref(), Some("rtk-uv"));
-        assert_eq!(dep.npm_name.as_deref(), Some("rtk-npm"));
-        assert!(dep.use_npx);
-        assert_eq!(dep.fallback.len(), 1);
-        assert_eq!(dep.fallback[0].method, "symlink");
-        assert_eq!(
-            dep.fallback[0].source.as_deref(),
-            Some("/usr/libexec/anolisa/tokenless/rtk")
-        );
-    }
-
-    #[test]
-    fn normalize_dep_null_fallback() {
-        let dep = normalize_dep(&json!(null));
-        assert_eq!(dep.binary, "");
-        assert_eq!(dep.package, "");
-        assert_eq!(dep.manager, "rpm");
-    }
-
-    #[test]
-    fn normalize_deps_mixed_array() {
-        let deps = normalize_deps(
-            &json!(["jq", "rtk>=0.35", {"binary": "curl", "package": "curl", "manager": "rpm"}]),
-        );
-        assert_eq!(deps.len(), 3);
-        assert_eq!(deps[0].binary, "jq");
-        assert_eq!(deps[0].manager, "rpm");
-        assert_eq!(deps[1].binary, "rtk");
-        assert_eq!(deps[1].version.as_deref(), Some(">=0.35"));
-        assert_eq!(deps[2].binary, "curl");
-        assert_eq!(deps[2].manager, "rpm");
-    }
-
-    #[test]
-    fn normalize_deps_empty() {
-        let deps = normalize_deps(&json!([]));
-        assert!(deps.is_empty());
-        let deps = normalize_deps(&json!(null));
-        assert!(deps.is_empty());
-    }
-
-    #[test]
-    fn extract_required_version_ge() {
-        assert_eq!(extract_required_version(">=0.35"), "0.35");
-    }
-
-    #[test]
-    fn extract_required_version_gt() {
-        assert_eq!(extract_required_version(">1.0"), "1.0");
-    }
-
-    #[test]
-    fn extract_required_version_no_operator() {
-        assert_eq!(extract_required_version("0.35"), "0.35");
-    }
-
-    #[test]
-    fn version_ge_equal() {
-        assert!(version_ge("0.35", "0.35"));
-    }
-
-    #[test]
-    fn version_ge_greater() {
-        assert!(version_ge("1.2.0", "1.0.0"));
-    }
-
-    #[test]
-    fn version_ge_less() {
-        assert!(!version_ge("0.34", "0.35"));
-    }
-
-    #[test]
-    fn version_ge_short_version() {
-        assert!(version_ge("2.0", "1.9.9"));
-    }
-
-    #[test]
-    fn version_ge_patch_comparison() {
-        assert!(version_ge("1.0.1", "1.0.0"));
-        assert!(!version_ge("1.0.0", "1.0.1"));
-    }
-
-    #[test]
-    fn build_json_result_ready() {
-        let result = build_json_result("Shell", &ReadyStatus::Ready, &[], &[]);
-        assert_eq!(result["tool"], "Shell");
-        assert_eq!(result["status"], "READY");
-        assert!(result.get("fixed").is_none());
-        assert!(result.get("missing").is_none());
-        assert!(result.get("diagnostic").is_none());
-    }
-
-    #[test]
-    fn build_json_result_not_ready() {
-        let result = build_json_result(
-            "Shell",
-            &ReadyStatus::NotReady,
-            &[],
-            &["fakebin99".to_string()],
-        );
-        assert_eq!(result["tool"], "Shell");
-        assert_eq!(result["status"], "NOT_READY");
-        assert_eq!(result["missing"][0], "fakebin99");
-        let diag = result["diagnostic"].as_str().unwrap();
-        assert!(diag.contains("Skip retry"));
-        assert!(diag.contains("required dependency missing"));
-    }
-
-    #[test]
-    fn build_json_result_unknown() {
-        let result = build_json_result("UnknownTool", &ReadyStatus::Unknown, &[], &[]);
-        assert_eq!(result["tool"], "UnknownTool");
-        assert_eq!(result["status"], "UNKNOWN");
-        assert!(result.get("fixed").is_none());
-        assert!(result.get("missing").is_none());
-        assert!(result.get("diagnostic").is_none());
-    }
-
-    #[test]
-    fn build_json_result_with_fixed() {
-        let result = build_json_result("Shell", &ReadyStatus::Ready, &["jq".to_string()], &[]);
-        assert_eq!(result["fixed"][0], "jq");
-    }
-
-    #[test]
-    fn format_status_all() {
-        assert_eq!(format_status(&ReadyStatus::Ready), "READY");
-        assert_eq!(format_status(&ReadyStatus::Partial), "PARTIAL");
-        assert_eq!(format_status(&ReadyStatus::NotReady), "NOT_READY");
-        assert_eq!(format_status(&ReadyStatus::Unknown), "UNKNOWN");
-    }
-
-    #[test]
-    fn format_dep_status_all() {
-        assert_eq!(format_dep_status(&DepStatus::Available), "✓");
-        assert_eq!(format_dep_status(&DepStatus::Missing), "missing");
-        let low = format_dep_status(&DepStatus::VersionLow {
-            installed: "0.34".to_string(),
-            required: "0.35".to_string(),
-        });
-        assert!(low.contains("0.34"));
-        assert!(low.contains("0.35"));
-    }
-
-    #[test]
-    fn expand_path_home() {
-        let home = std::env::var("HOME").unwrap_or_else(|_| "/tmp".to_string());
-        let expanded = expand_path("~/.copilot-shell/settings.json");
-        assert_eq!(expanded, format!("{}/.copilot-shell/settings.json", home));
-    }
-
-    #[test]
-    fn expand_path_absolute() {
-        let expanded = expand_path("/etc/config.json");
-        assert_eq!(expanded, "/etc/config.json");
-    }
-
-    #[test]
-    fn version_ge_prefixed_v() {
-        assert!(version_ge("v22.1.0", "16.0.0"));
-        assert!(version_ge("V22.1.0", "16.0.0"));
-    }
-
-    #[test]
-    fn version_ge_build_suffix() {
-        assert!(version_ge("1.2.3-rc1", "1.2.0"));
-        assert!(version_ge("1.2.3+build", "1.2.3"));
-    }
-
-    #[test]
-    fn version_ge_short_segments() {
-        assert!(version_ge("22.1", "16.0"));
-        assert!(!version_ge("1.0", "2.0"));
-    }
-
-    #[test]
-    fn load_spec_skips_meta_keys() {
-        let tmp_dir = std::env::temp_dir();
-        let spec_path = tmp_dir.join("test-tool-ready-spec.json");
-        let spec_content = json!({
-            "_meta": {"version": "2.0"},
-            "_comment": "this should be skipped",
-            "Shell": {
-                "required": ["jq"],
-                "recommended": [],
-                "config_files": [],
-                "permissions": [],
-                "network": []
-            }
-        });
-        std::fs::write(&spec_path, serde_json::to_string(&spec_content).unwrap()).unwrap();
-
-        let specs = load_spec(&spec_path).unwrap();
-        assert!(!specs.contains_key("_meta"));
-        assert!(!specs.contains_key("_comment"));
-        assert!(specs.contains_key("Shell"));
-        let shell_spec = specs.get("Shell").unwrap();
-        assert_eq!(shell_spec.required.len(), 1);
-        assert_eq!(shell_spec.required[0].binary, "jq");
-
-        std::fs::remove_file(&spec_path).ok();
-    }
-
-    #[test]
-    fn load_spec_mixed_formats() {
-        let tmp_dir = std::env::temp_dir();
-        let spec_path = tmp_dir.join("test-mixed-spec.json");
-        let spec_content = json!({
-            "Shell": {
-                "required": ["jq", "rtk>=0.35", {"binary": "curl", "package": "curl", "manager": "rpm"}],
-                "recommended": [],
-                "config_files": [],
-                "permissions": [],
-                "network": []
-            }
-        });
-        std::fs::write(&spec_path, serde_json::to_string(&spec_content).unwrap()).unwrap();
-
-        let specs = load_spec(&spec_path).unwrap();
-        let shell_spec = specs.get("Shell").unwrap();
-        assert_eq!(shell_spec.required.len(), 3);
-        assert_eq!(shell_spec.required[0].binary, "jq");
-        assert_eq!(shell_spec.required[0].manager, "rpm");
-        assert_eq!(shell_spec.required[1].binary, "rtk");
-        assert_eq!(shell_spec.required[1].version.as_deref(), Some(">=0.35"));
-        assert_eq!(shell_spec.required[2].binary, "curl");
-        assert_eq!(shell_spec.required[2].manager, "rpm");
-
-        std::fs::remove_file(&spec_path).ok();
-    }
-
-    #[cfg(unix)]
-    fn make_test_dir(label: &str) -> std::path::PathBuf {
-        use std::time::{SystemTime, UNIX_EPOCH};
-        let nanos = SystemTime::now()
-            .duration_since(UNIX_EPOCH)
-            .unwrap()
-            .as_nanos();
-        let p = std::env::temp_dir().join(format!(
-            "tokenless-is-trusted-{}-{}-{}",
-            std::process::id(),
-            nanos,
-            label
-        ));
-        let _ = std::fs::remove_dir_all(&p);
-        std::fs::create_dir_all(&p).unwrap();
-        p
-    }
-
-    #[cfg(unix)]
-    fn chmod_file(path: &std::path::Path, mode: u32) {
-        use std::os::unix::fs::PermissionsExt;
-        let mut perm = std::fs::metadata(path).unwrap().permissions();
-        perm.set_mode(mode);
-        std::fs::set_permissions(path, perm).unwrap();
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn is_trusted_path_system_prefixes_unconditional() {
-        // The system-path branch returns early without touching the
-        // filesystem, so non-existent paths still report trusted.
-        use std::path::Path;
-        assert!(is_trusted_path(Path::new("/usr/share/anolisa/x")));
-        assert!(is_trusted_path(Path::new("/usr/libexec/anolisa/x")));
-        assert!(is_trusted_path(Path::new("/usr/lib/anolisa/x")));
-        assert!(is_trusted_path(Path::new("/usr/local/share/anolisa/x")));
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn is_trusted_path_rejects_world_writable_parent() {
-        use std::os::unix::fs::MetadataExt;
-        let tmp = make_test_dir("ww-parent");
-        if std::fs::metadata(&tmp).unwrap().uid() != current_uid() {
-            // /tmp on hardened multi-user systems may strip our ownership;
-            // the world-writable check is moot in that case.
-            std::fs::remove_dir_all(&tmp).ok();
-            return;
-        }
-        chmod_file(&tmp, 0o777);
-        let f = tmp.join("binary");
-        std::fs::write(&f, b"#!/bin/sh\n").unwrap();
-        chmod_file(&f, 0o755);
-        assert!(
-            !is_trusted_path(&f),
-            "world-writable parent dir must be rejected"
-        );
-        chmod_file(&tmp, 0o755);
-        std::fs::remove_dir_all(&tmp).ok();
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn is_trusted_path_rejects_world_writable_file() {
-        use std::os::unix::fs::MetadataExt;
-        let tmp = make_test_dir("ww-file");
-        if std::fs::metadata(&tmp).unwrap().uid() != current_uid() {
-            std::fs::remove_dir_all(&tmp).ok();
-            return;
-        }
-        chmod_file(&tmp, 0o755);
-        let f = tmp.join("binary");
-        std::fs::write(&f, b"#!/bin/sh\n").unwrap();
-        chmod_file(&f, 0o777);
-        assert!(
-            !is_trusted_path(&f),
-            "world-writable file mode must be rejected"
-        );
-        std::fs::remove_dir_all(&tmp).ok();
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn is_trusted_path_accepts_owned_safe_file() {
-        use std::os::unix::fs::MetadataExt;
-        let tmp = make_test_dir("ok");
-        if std::fs::metadata(&tmp).unwrap().uid() != current_uid() {
-            std::fs::remove_dir_all(&tmp).ok();
-            return;
-        }
-        chmod_file(&tmp, 0o755);
-        let f = tmp.join("binary");
-        std::fs::write(&f, b"#!/bin/sh\n").unwrap();
-        chmod_file(&f, 0o755);
-        assert!(
-            is_trusted_path(&f),
-            "uid-owned non-writable file must be accepted"
-        );
-        std::fs::remove_dir_all(&tmp).ok();
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn is_trusted_path_rejects_nonexistent_file() {
-        let nonexistent = std::env::temp_dir().join(format!(
-            "tokenless-nonexistent-{}-{}",
-            std::process::id(),
-            std::time::SystemTime::now()
-                .duration_since(std::time::UNIX_EPOCH)
-                .unwrap()
-                .as_nanos()
-        ));
-        assert!(
-            !is_trusted_path(&nonexistent),
-            "non-existent file must be rejected"
-        );
-    }
-
-    #[test]
-    fn expand_path_rejects_parent_dir_traversal() {
-        // ParentDir components in ~/... paths are rejected at the syntax
-        // layer so a misconfigured config_files entry like "~/../etc/passwd"
-        // cannot escape the home directory after expansion.
-        let escaped = expand_path("~/../etc/passwd");
-        assert_eq!(
-            escaped, "~/../etc/passwd",
-            "ParentDir-bearing tilde path must be returned unchanged"
-        );
-        let escaped2 = expand_path("~/sub/../../../etc/passwd");
-        assert_eq!(
-            escaped2, "~/sub/../../../etc/passwd",
-            "Deep ParentDir traversal must be returned unchanged"
-        );
-    }
-
-    #[test]
-    fn generate_checklist_unknown_status() {
-        let results = [ToolReadyResult {
-            tool_name: "UnknownTool".to_string(),
-            status: ReadyStatus::Unknown,
-            required_results: vec![(
-                DepEntry {
-                    binary: "fake".to_string(),
-                    version: None,
-                    package: "fake".to_string(),
-                    apt_package: None,
-                    apk_package: None,
-                    manager: "rpm".to_string(),
-                    pip_name: None,
-                    uv_name: None,
-                    npm_name: None,
-                    use_npx: false,
-                    fallback: vec![],
-                },
-                DepStatus::Missing,
-            )],
-            recommended_results: vec![],
-            config_results: vec![],
-            permission_results: vec![],
-            network_results: vec![],
-        }];
-        let checklist = generate_checklist(&results);
-        assert!(checklist.contains("UNKNOWN"));
-        assert!(checklist.contains("unknown"));
-    }
+    include!("tests/env_check_tests.rs");
 }
