+++ src/tokenless/crates/tokenless-cli/src/main.rs
@@ -927,106 +927,5 @@ fn main() {
 #[cfg(test)]
 mod tests {
     use super::*;
-
-    fn temp_subdir(label: &str) -> std::path::PathBuf {
-        use std::time::{SystemTime, UNIX_EPOCH};
-        let nanos = SystemTime::now()
-            .duration_since(UNIX_EPOCH)
-            .unwrap()
-            .as_nanos();
-        let p = std::env::temp_dir().join(format!(
-            "tokenless-db-validate-{}-{}-{}",
-            std::process::id(),
-            nanos,
-            label
-        ));
-        std::fs::create_dir_all(&p).unwrap();
-        p
-    }
-
-    #[test]
-    fn validate_db_path_rejects_empty_home() {
-        // No trusted home anchor means starts_with("") would match
-        // any path, so the function must short-circuit to rejection.
-        let err = validate_db_path("/tmp/whatever.db", "").unwrap_err();
-        assert!(err.contains("no trusted home"));
-    }
-
-    #[test]
-    fn validate_db_path_accepts_path_inside_home() {
-        let home = temp_subdir("inside");
-        let canon_home = std::fs::canonicalize(&home).unwrap();
-        let inner = canon_home.join("stats.db");
-        let result =
-            validate_db_path(inner.to_str().unwrap(), canon_home.to_str().unwrap()).unwrap();
-        assert_eq!(result, inner.to_str().unwrap());
-        std::fs::remove_dir_all(&home).ok();
-    }
-
-    #[test]
-    fn validate_db_path_rejects_path_outside_home() {
-        let home = temp_subdir("outside-home");
-        let canon_home = std::fs::canonicalize(&home).unwrap();
-        // Pick a known-existing directory that is NOT under home.
-        let outside = std::path::Path::new("/etc/hosts");
-        if !outside.exists() {
-            std::fs::remove_dir_all(&home).ok();
-            return;
-        }
-        let err = validate_db_path("/etc/hosts", canon_home.to_str().unwrap()).unwrap_err();
-        assert!(err.contains("outside home"));
-        std::fs::remove_dir_all(&home).ok();
-    }
-
-    #[test]
-    fn validate_db_path_rejects_parent_dir_bypass_with_existing_parent() {
-        // ~/foo/../../etc/evil.db where /etc exists: canonicalize() of
-        // the parent resolves to /etc, which must fail starts_with(home).
-        let home = temp_subdir("pd-existing");
-        let canon_home = std::fs::canonicalize(&home).unwrap();
-        let escape = canon_home.join("foo/../../etc/evil.db");
-        let err =
-            validate_db_path(escape.to_str().unwrap(), canon_home.to_str().unwrap()).unwrap_err();
-        // Either "outside home" (parent canonicalized away from home) or
-        // "cannot be resolved" (parent itself unreachable). Both are valid
-        // rejections — what matters is no Ok return.
-        assert!(err.contains("outside home") || err.contains("cannot be resolved"));
-        std::fs::remove_dir_all(&home).ok();
-    }
-
-    #[test]
-    fn validate_db_path_canonicalizes_home_with_symlink_prefix() {
-        // If the caller passes a home that contains a symlink in any
-        // prefix (e.g. /tmp on macOS resolves to /private/tmp), the
-        // candidate path will canonicalize to the resolved form and
-        // diverge from the raw home unless validate_db_path canonicalizes
-        // home too. Linux /tmp has no such symlink, so the assertion is
-        // informational there but real coverage on macOS.
-        let home = temp_subdir("sym-prefix");
-        let inner = home.join("stats.db");
-        let result = validate_db_path(inner.to_str().unwrap(), home.to_str().unwrap());
-        assert!(
-            result.is_ok(),
-            "raw (non-canonical) home should be accepted after internal canonicalization: {:?}",
-            result
-        );
-        std::fs::remove_dir_all(&home).ok();
-    }
-
-    #[test]
-    fn validate_db_path_rejects_parent_dir_bypass_with_nonexistent_parent() {
-        // ~/nonexistent-path/../../etc/evil.db where nonexistent-path
-        // doesn't exist: parent canonicalize() ALSO fails, so without the
-        // hardening this path would slip through via the old fallback.
-        let home = temp_subdir("pd-nonexistent");
-        let canon_home = std::fs::canonicalize(&home).unwrap();
-        let escape = canon_home.join("does-not-exist-xyz/../../etc/evil.db");
-        let result = validate_db_path(escape.to_str().unwrap(), canon_home.to_str().unwrap());
-        assert!(
-            result.is_err(),
-            "ParentDir bypass via nonexistent intermediate must be rejected; got {:?}",
-            result
-        );
-        std::fs::remove_dir_all(&home).ok();
-    }
+    include!("tests/main_tests.rs");
 }
