+++ src/tokenless/crates/tokenless-cli/src/mcp.rs
@@ -191,83 +191,5 @@ fn tool_error(message: &str) -> Value {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use tokenless_ccr::InMemoryStore;
-
-    #[test]
-    fn tool_list_exposes_retrieve() {
-        let list = retrieve_tool();
-        assert_eq!(list["name"], json!("tokenless_retrieve"));
-        assert!(list["inputSchema"]["properties"]["hash"]["type"].is_string());
-    }
-
-    #[test]
-    fn dispatch_unknown_tool_is_error() {
-        let r = handle_tool_call(json!({"name":"nope","arguments":{}}), &None);
-        assert_eq!(r["isError"], json!(true));
-        assert!(
-            r["content"][0]["text"]
-                .as_str()
-                .unwrap()
-                .contains("unknown tool")
-        );
-    }
-
-    #[test]
-    fn retrieve_missing_hash_is_error() {
-        let r = retrieve(json!({}), &None);
-        assert_eq!(r["isError"], json!(true));
-        assert!(
-            r["content"][0]["text"]
-                .as_str()
-                .unwrap()
-                .contains("missing")
-        );
-    }
-
-    #[test]
-    fn retrieve_invalid_hash_format_is_error() {
-        // A non-hash argument (e.g. a file path) gets a format error before
-        // any DB round-trip, not a misleading "no stashed payload".
-        let r = retrieve(json!({"hash": "/some/path"}), &None);
-        assert_eq!(r["isError"], json!(true));
-        assert!(
-            r["content"][0]["text"]
-                .as_str()
-                .unwrap()
-                .contains("invalid stash hash")
-        );
-    }
-
-    #[test]
-    fn retrieve_round_trips_via_store() {
-        let store: Arc<dyn StashStore> = Arc::new(InMemoryStore::new());
-        let key = store.stash("payload-body").unwrap();
-        let r = retrieve_from_store(&store, &key);
-        assert_eq!(r["isError"], json!(false));
-        assert_eq!(r["content"][0]["text"], json!("payload-body"));
-    }
-
-    #[test]
-    fn retrieve_accepts_marker_line() {
-        let store: Arc<dyn StashStore> = Arc::new(InMemoryStore::new());
-        let key = store.stash("dropped items").unwrap();
-        let marker_line = format!("<... 5 items truncated, retrieve with <<tokenless:{key}>>");
-        // Exercise the public entry point so marker extraction in `retrieve`
-        // is covered, not just the bare-key path of `retrieve_from_store`.
-        let r = retrieve(json!({"hash": marker_line}), &Some(store));
-        assert_eq!(r["content"][0]["text"], json!("dropped items"));
-    }
-
-    #[test]
-    fn retrieve_missing_payload_is_error() {
-        let store: Arc<dyn StashStore> = Arc::new(InMemoryStore::new());
-        let r = retrieve_from_store(&store, "000000000000000000000000");
-        assert_eq!(r["isError"], json!(true));
-        assert!(
-            r["content"][0]["text"]
-                .as_str()
-                .unwrap()
-                .contains("no stashed payload")
-        );
-    }
+    include!("tests/mcp_tests.rs");
 }
