+++ src/tokenless/crates/tokenless-stats/src/record.rs
@@ -286,125 +286,5 @@ impl StatsRecord {
 #[cfg(test)]
 mod tests {
     use super::*;
-
-    #[test]
-    fn test_operation_type_from_str() {
-        assert_eq!(
-            OperationType::from_str("compress-schema").unwrap(),
-            OperationType::CompressSchema
-        );
-        assert_eq!(
-            OperationType::from_str("compress-response").unwrap(),
-            OperationType::CompressResponse
-        );
-        assert_eq!(
-            OperationType::from_str("rewrite-command").unwrap(),
-            OperationType::RewriteCommand
-        );
-        assert_eq!(
-            OperationType::from_str("compress-toon").unwrap(),
-            OperationType::CompressToon
-        );
-        assert!(OperationType::from_str("unknown").is_err());
-    }
-
-    #[test]
-    fn test_savings_calculation() {
-        let record = StatsRecord::new(
-            OperationType::CompressSchema,
-            "copilot-shell".to_string(),
-            1000,
-            400,
-            500,
-            200,
-        );
-
-        assert_eq!(record.chars_saved(), 500);
-        assert_eq!(record.tokens_saved(), 200);
-        assert!((record.chars_percent() - 50.0).abs() < 0.1);
-        assert!((record.tokens_percent() - 50.0).abs() < 0.1);
-    }
-
-    #[test]
-    fn test_record_with_text() {
-        let record = StatsRecord::new(
-            OperationType::CompressSchema,
-            "copilot-shell".to_string(),
-            16,
-            4,
-            10,
-            3,
-        )
-        .with_text("original text here".to_string(), "compressed".to_string());
-
-        assert!(record.before_text.is_some());
-        assert!(record.after_text.is_some());
-    }
-
-    #[test]
-    fn test_format_summary_line() {
-        let record = StatsRecord::new(
-            OperationType::CompressSchema,
-            "copilot-shell".to_string(),
-            1000,
-            400,
-            500,
-            200,
-        )
-        .with_session_id("session-123")
-        .with_tool_use_id("call_abc");
-
-        let line = record.format_summary_line();
-        assert!(line.contains("[ID:-1]"));
-        assert!(line.contains("copilot-shell"));
-        assert!(line.contains("Session:session-123"));
-        assert!(line.contains("Tool:call_abc"));
-    }
-
-    #[test]
-    fn test_format_summary_line_with_pid() {
-        let record = StatsRecord::new(
-            OperationType::CompressSchema,
-            "copilot-shell".to_string(),
-            1000,
-            400,
-            500,
-            200,
-        )
-        .with_session_id("session-123")
-        .with_source_pid(12345);
-
-        let line = record.format_summary_line();
-        assert!(line.contains("copilot-shell"));
-        assert!(line.contains("pid:12345"));
-    }
-
-    #[test]
-    fn test_compression_mode_roundtrip() {
-        assert_eq!(CompressionMode::Active.as_str(), "active");
-        assert_eq!(CompressionMode::DryRun.as_str(), "dry-run");
-        assert_eq!(CompressionMode::from_db("active"), CompressionMode::Active);
-        assert_eq!(CompressionMode::from_db("dry-run"), CompressionMode::DryRun);
-        // Legacy "dryrun" form still readable (backward compatibility)
-        assert_eq!(CompressionMode::from_db("dryrun"), CompressionMode::DryRun);
-        // Unknown / empty (legacy NULL) fall back to Active
-        assert_eq!(CompressionMode::from_db(""), CompressionMode::Active);
-        assert_eq!(CompressionMode::from_db("???"), CompressionMode::Active);
-    }
-
-    #[test]
-    fn test_record_default_mode_active() {
-        let record = StatsRecord::new(
-            OperationType::CompressSchema,
-            "test".to_string(),
-            100,
-            25,
-            80,
-            20,
-        );
-        assert_eq!(record.mode, CompressionMode::Active);
-
-        let record = record.with_mode(CompressionMode::DryRun);
-        assert_eq!(record.mode, CompressionMode::DryRun);
-    }
+    include!("tests/record_tests.rs");
 }
