+++ src/tokenless/crates/tokenless-stats/src/recorder.rs
@@ -383,108 +383,5 @@ impl StatsSummary {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::record::OperationType;
-
-    fn new_recorder() -> (StatsRecorder, tempfile::TempDir) {
-        let dir = tempfile::tempdir().unwrap();
-        let db = dir.path().join("stats.db");
-        let rec = StatsRecorder::new(&db).unwrap();
-        (rec, dir)
-    }
-
-    fn sample(op: OperationType, mode: CompressionMode, session: &str) -> StatsRecord {
-        StatsRecord::new(op, "cli".to_string(), 1000, 400, 500, 200)
-            .with_session_id(session)
-            .with_mode(mode)
-    }
-
-    #[test]
-    fn records_and_reads_mode() {
-        let (rec, _dir) = new_recorder();
-        let id = rec
-            .record(&sample(
-                OperationType::CompressSchema,
-                CompressionMode::DryRun,
-                "s1",
-            ))
-            .unwrap();
-        let got = rec.record_by_id(id).unwrap().unwrap();
-        assert_eq!(got.mode, CompressionMode::DryRun);
-        assert_eq!(got.session_id.as_deref(), Some("s1"));
-    }
-
-    #[test]
-    fn records_and_reads_stash_fields() {
-        let (rec, _dir) = new_recorder();
-        let rec_in = sample(
-            OperationType::CompressResponse,
-            CompressionMode::Active,
-            "stash-ses",
-        )
-        .with_stash(Some(3), Some(0), Some(42));
-        let id = rec.record(&rec_in).unwrap();
-        let got = rec.record_by_id(id).unwrap().unwrap();
-        assert_eq!(got.stash_writes, Some(3));
-        assert_eq!(got.stash_errors, Some(0));
-        assert_eq!(got.stash_size, Some(42));
-    }
-
-    #[test]
-    fn stash_fields_default_none_when_unstashed() {
-        let (rec, _dir) = new_recorder();
-        let id = rec
-            .record(&sample(
-                OperationType::CompressResponse,
-                CompressionMode::Active,
-                "no-stash",
-            ))
-            .unwrap();
-        let got = rec.record_by_id(id).unwrap().unwrap();
-        assert_eq!(got.stash_writes, None);
-        assert_eq!(got.stash_errors, None);
-        assert_eq!(got.stash_size, None);
-    }
-
-    #[test]
-    fn default_mode_is_active() {
-        let (rec, _dir) = new_recorder();
-        let id = rec
-            .record(&sample(
-                OperationType::CompressSchema,
-                CompressionMode::Active,
-                "s1",
-            ))
-            .unwrap();
-        let got = rec.record_by_id(id).unwrap().unwrap();
-        assert_eq!(got.mode, CompressionMode::Active);
-    }
-
-    #[test]
-    fn records_by_session_filters() {
-        let (rec, _dir) = new_recorder();
-        rec.record(&sample(
-            OperationType::CompressResponse,
-            CompressionMode::Active,
-            "baseline",
-        ))
-        .unwrap();
-        rec.record(&sample(
-            OperationType::CompressResponse,
-            CompressionMode::DryRun,
-            "tokenless",
-        ))
-        .unwrap();
-        rec.record(&sample(
-            OperationType::CompressResponse,
-            CompressionMode::Active,
-            "baseline",
-        ))
-        .unwrap();
-
-        let baseline = rec.records_by_session("baseline", None).unwrap();
-        let tokenless = rec.records_by_session("tokenless", None).unwrap();
-        assert_eq!(baseline.len(), 2);
-        assert_eq!(tokenless.len(), 1);
-        assert_eq!(tokenless[0].mode, CompressionMode::DryRun);
-    }
+    include!("tests/recorder_tests.rs");
 }
